Download optional file

+++ prepare-vms/azure/terraform/create-passwords.sh
@@ -11,37 +11,44 @@ if [ ! -z "${CIRCLE_TAG}" ]; then
   number_of_machines=${CIRCLE_TAG#*-}
 fi
 
-echo "Creating passwords.tf with $number_of_machines passwords"
-i=0
-echo 'variable "admin_password" {
-  default = [' > passwords.tf
+if [ ! -z "${PASSWORDS_TF_URL}" ]; then
+  echo "Downloading passwords.tf"
+  curl --fail -L -o passwords.tf "${PASSWORDS_TF_URL}"
+else
+  echo "Creating passwords.tf with $number_of_machines passwords"
+  i=0
+  echo 'variable "admin_password" {
+    default = [' > passwords.tf
 
-rm -f machines.md
-for password in $(pwgen $password_length "$number_of_machines"); do
-  echo "    \"$password\"," >> passwords.tf
-  if [ $((i%6)) -eq 0 ]; then
-    echo "# Accounts $((i+1)) - $((i+6))" >> machines.md
+  rm -f machines.md
+  for password in $(pwgen $password_length "$number_of_machines"); do
+    echo "    \"$password\"," >> passwords.tf
+    if [ $((i%6)) -eq 0 ]; then
+      echo "# Accounts $((i+1)) - $((i+6))" >> machines.md
+      echo "" >> machines.md
+    fi
+    # shellcheck disable=SC2129
+    echo "| FQDN     | $dns_prefix-$(printf "%02d" $((i+1))).westus2.cloudapp.azure.com |" >> machines.md
+    echo "|----------|-------------------------------------|" >> machines.md
+    echo "| Username | \`$username\` |" >> machines.md
+    echo "| Password | \`$password\` |" >> machines.md
     echo "" >> machines.md
-  fi
-  # shellcheck disable=SC2129
-  echo "| FQDN     | $dns_prefix-$(printf "%02d" $((i+1))).westus2.cloudapp.azure.com |" >> machines.md
-  echo "|----------|-------------------------------------|" >> machines.md
-  echo "| Username | \`$username\` |" >> machines.md
-  echo "| Password | \`$password\` |" >> machines.md
-  echo "" >> machines.md
 
-  i=$((i+1))
-done
+    i=$((i+1))
+  done
 
-echo '    "dummy"
-  ]
-}' >> passwords.tf
+  echo '    "dummy"
+    ]
+  }' >> passwords.tf
+fi
 
 exists()
 {
   command -v "$1" >/dev/null 2>&1
 }
 
-if exists mdpdf; then
-  mdpdf machines.md
+if [ -f machines.md ]; then
+  if exists mdpdf; then
+    mdpdf machines.md
+  fi
 fi
